Rename package, consistently specify nullity
