+++ src/main/java/me/lokka30/treasury/plugin/command/treasury/subcommand/migrate/FailureConsumer.java
@@ -1,4 +1,4 @@
-package me.lokka30.treasury.plugin.command.treasury.subcommand.migration;
+package me.lokka30.treasury.plugin.command.treasury.subcommand.migrate;
 
 import me.lokka30.treasury.api.economy.response.EconomyException;
 import org.jetbrains.annotations.NotNull;
@@ -7,7 +7,7 @@ import java.util.concurrent.Phaser;
 import java.util.function.Consumer;
 
 final class FailureConsumer<T> extends PhasedSubscriber<T> {
-    private final Consumer<EconomyException> consumer;
+    private final @NotNull Consumer<EconomyException> consumer;
 
     FailureConsumer(@NotNull Phaser phaser, @NotNull Consumer<EconomyException> consumer) {
         super(phaser);
