+++ src/main/java/me/lokka30/treasury/plugin/command/treasury/subcommand/migrate/TriConsumer.java
@@ -1,0 +1,7 @@
+package me.lokka30.treasury.plugin.command.treasury.subcommand.migrate;
+
+interface TriConsumer<T, U, V> {
+
+    void accept(T t, U u, V v);
+
+}
