--- src/main/java/me/lokka30/treasury/plugin/command/treasury/subcommand/migration/TriConsumer.java
@@ -1,7 +1,0 @@
-package me.lokka30.treasury.plugin.command.treasury.subcommand.migration;
-
-interface TriConsumer<T, U, V> {
-
-    void accept(T t, U u, V v);
-
-}
